Move `ActionResult` into own file
