+++ frontend/src/shared/lib/api/actionResult.ts
@@ -1,0 +1,1 @@
+export type ActionResult<T> = { success: true; data: T } | { success: false; error: string };
