+++ frontend/src/shared/types/course.ts
@@ -144,5 +144,3 @@ export interface Page<T> {
   last: boolean;
   empty: boolean;
 }
-
-export type ActionResult<T> = { success: true; data: T } | { success: false; error: string };
